MondainLegacy
- Moved ElfOrHuman to RaceDefintions.cs
- Moved, and seperated what was actually a duplicate command and put it in Quests.
