+++ Scripts/Items/Artifacts/Equipment/Armor/AegisOfGrace.cs
@@ -1,3 +1,5 @@
+using Server.Misc;
+
 namespace Server.Items
 {
     public class AegisOfGrace : DragonHelm, ICanBeElfOrHuman
