+++ Scripts/Items/Artifacts/Equipment/Armor/HelmOfSwiftness.cs
@@ -1,3 +1,5 @@
+using Server.Misc;
+
 namespace Server.Items
 {
     public class HelmOfSwiftness : WingedHelm, ICanBeElfOrHuman
