+++ Scripts/Items/Artifacts/Equipment/Clothing/CloakOfDeath.cs
@@ -1,3 +1,5 @@
+using Server.Misc;
+
 namespace Server.Items
 {
     [Flipable(0x2FB9, 0x3173)]
