+++ Scripts/Items/Artifacts/Equipment/Clothing/RobeOfTheEquinox.cs
@@ -1,3 +1,5 @@
+using Server.Misc;
+
 namespace Server.Items
 {
     [Flipable(0x1F03, 0x1F04)]
